+++ vector/src/main/java/im/vector/riotx/features/roomprofile/members/RoomMemberListController.kt
@@ -17,7 +17,11 @@
 package im.vector.riotx.features.roomprofile.members
 
 import com.airbnb.epoxy.TypedEpoxyController
+import im.vector.matrix.android.api.session.events.model.Event
+import im.vector.matrix.android.api.session.events.model.toModel
 import im.vector.matrix.android.api.session.room.model.RoomMemberSummary
+import im.vector.matrix.android.api.session.room.model.RoomThirdPartyInviteContent
+import im.vector.matrix.android.api.util.MatrixItem
 import im.vector.matrix.android.api.util.toMatrixItem
 import im.vector.riotx.R
 import im.vector.riotx.core.epoxy.dividerItem
@@ -37,6 +41,7 @@ class RoomMemberListController @Inject constructor(
 
     interface Callback {
         fun onRoomMemberClicked(roomMember: RoomMemberSummary)
+        fun onThreePidInvites(event: Event)
     }
 
     private val dividerColor = colorProvider.getColorFromAttribute(R.attr.vctr_list_divider_color)
@@ -76,5 +81,34 @@ class RoomMemberListController @Inject constructor(
                     }
             )
         }
+        buildThreePidInvites(data)
+    }
+
+    private fun buildThreePidInvites(data: RoomMemberListViewState) {
+        if (data.threePidInvites().isNullOrEmpty()) {
+            return
+        }
+
+        buildProfileSection(
+                stringProvider.getString(R.string.room_member_power_level_three_pid_invites)
+        )
+
+        data.threePidInvites()?.forEachIndexed { idx, event ->
+            val content = event.content.toModel<RoomThirdPartyInviteContent>() ?: return@forEachIndexed
+
+            profileMatrixItem {
+                id("3pid_$idx")
+                matrixItem(content.toMatrixItem())
+                avatarRenderer(avatarRenderer)
+                clickListener { _ ->
+                    callback?.onThreePidInvites(event)
+                }
+            }
+
+        }
+    }
+
+    private fun RoomThirdPartyInviteContent.toMatrixItem(): MatrixItem {
+        return MatrixItem.UserItem("@", displayName = displayName)
     }
 }