+++ vector/src/main/java/im/vector/riotx/features/roomprofile/members/RoomMemberListFragment.kt
@@ -23,6 +23,7 @@ import android.view.View
 import com.airbnb.mvrx.args
 import com.airbnb.mvrx.fragmentViewModel
 import com.airbnb.mvrx.withState
+import im.vector.matrix.android.api.session.events.model.Event
 import im.vector.matrix.android.api.session.room.model.RoomMemberSummary
 import im.vector.matrix.android.api.util.toMatrixItem
 import im.vector.riotx.R
@@ -86,6 +87,10 @@ class RoomMemberListFragment @Inject constructor(
 
     override fun onRoomMemberClicked(roomMember: RoomMemberSummary) {
         navigator.openRoomMemberProfile(roomMember.userId, roomId = roomProfileArgs.roomId, context = requireActivity())
+    }
+
+    override fun onThreePidInvites(event: Event) {
+        // TODO Display a bottom sheet to revoke invite if power level is high enough
     }
 
     private fun renderRoomSummary(state: RoomMemberListViewState) {